git-svn-id: https://svn.kuali.org/repos/student/trunk@17802 0ac02d80-fb3f-11dc-a574-c57244a850eb

+++ ks-core/ks-core-ui/src/main/java/org/kuali/student/core/organization/assembly/OrgOrgRelationAssembler.java
@@ -24,18 +24,20 @@ import static org.kuali.student.core.assembly.util.AssemblerUtils.isModified;
 import static org.kuali.student.core.assembly.util.AssemblerUtils.isUpdated;
 
 import java.util.ArrayList;
+import java.util.Iterator;
 import java.util.List;
 
 import org.apache.log4j.Logger;
 import org.kuali.student.common.ui.client.mvc.DataModel;
 import org.kuali.student.common.ui.client.mvc.DataModelDefinition;
-import org.kuali.student.core.assembly.Assembler;
 import org.kuali.student.core.assembly.data.AssemblyException;
 import org.kuali.student.core.assembly.data.Data;
+import org.kuali.student.core.assembly.data.Data.Property;
 import org.kuali.student.core.assembly.data.Metadata;
 import org.kuali.student.core.assembly.data.QueryPath;
-import org.kuali.student.core.assembly.data.SaveResult;
-import org.kuali.student.core.assembly.data.Data.Property;
+import org.kuali.student.core.assembly.old.Assembler;
+import org.kuali.student.core.assembly.old.data.SaveResult;
+import org.kuali.student.core.assembly.util.AssemblerUtils;
 import org.kuali.student.core.dto.MetaInfo;
 import org.kuali.student.core.dto.StatusInfo;
 import org.kuali.student.core.exceptions.DoesNotExistException;
@@ -160,7 +162,8 @@ public class OrgOrgRelationAssembler implements Assembler<Data, OrgorgRelationHe
         orgOrgModel.setDefinition(def);
         QueryPath metaPath = QueryPath.concat(null, ORGORG_PATH);
         Metadata orgOrgMeta =orgOrgModel.getMetadata(metaPath);
-        for (Property p : (Data)input.get("orgOrgRelationInfo")) {
+        for (Iterator<Property> propIter = ((Data)input.get("orgOrgRelationInfo")).iterator();propIter.hasNext();) {
+        	Property p = propIter.next();
             OrgorgRelationHelper orgOrgRelation=  OrgorgRelationHelper.wrap((Data)p.getValue());
             if (isUpdated(orgOrgRelation.getData())) {
                 if (orgOrgMeta.isCanEdit()) {
@@ -174,13 +177,15 @@ public class OrgOrgRelationAssembler implements Assembler<Data, OrgorgRelationHe
                         throw (new AssemblyException());
                     }
                 }
+                AssemblerUtils.setUpdated(orgOrgRelation.getData(), false);
             }
-            else if(isDeleted(orgOrgRelation.getData())){
+            else if(isDeleted(orgOrgRelation.getData())&&orgOrgRelation.getId()!=null){
 //              OrgOrgRelationInfo orgOrgRelationInfo = buildOrgOrgRelationInfo(orgOrgRelation);
 //              orgOrgRelationInfo.setId(orgOrgRelation.getId());
               try{
                   if(orgOrgRelation.getId()!=null){
                       StatusInfo  result = orgService.removeOrgOrgRelation(orgOrgRelation.getId());
+                      propIter.remove();
                   }
               }
               catch(Exception e ){
@@ -202,6 +207,7 @@ public class OrgOrgRelationAssembler implements Assembler<Data, OrgorgRelationHe
                     LOG.error(e);
                     throw(new AssemblyException());
                 }
+                AssemblerUtils.setCreated(orgOrgRelation.getData(), false);
             }
